--- app/src/main/res/drawable/ic_heart_outline.xml
@@ -1,4 +1,0 @@
-<vector android:height="32dp" android:viewportHeight="512"
-    android:viewportWidth="512" android:width="32dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="#f50000" android:pathData="M225.8,468.2l-2.5,-2.3L48.1,303.2C17.4,274.7 0,234.7 0,192.8v-3.3c0,-70.4 50,-130.8 119.2,-144C158.6,37.9 198.9,47 231,69.6c9,6.4 17.4,13.8 25,22.3c4.2,-4.8 8.7,-9.2 13.5,-13.3c3.7,-3.2 7.5,-6.2 11.5,-9c0,0 0,0 0,0C313.1,47 353.4,37.9 392.8,45.4C462,58.6 512,119.1 512,189.5v3.3c0,41.9 -17.4,81.9 -48.1,110.4L288.7,465.9l-2.5,2.3c-8.2,7.6 -19,11.9 -30.2,11.9s-22,-4.2 -30.2,-11.9zM239.1,145c-0.4,-0.3 -0.7,-0.7 -1,-1.1l-17.8,-20c0,0 -0.1,-0.1 -0.1,-0.1c0,0 0,0 0,0c-23.1,-25.9 -58,-37.7 -92,-31.2C81.6,101.5 48,142.1 48,189.5v3.3c0,28.5 11.9,55.8 32.8,75.2L256,430.7 431.2,268c20.9,-19.4 32.8,-46.7 32.8,-75.2v-3.3c0,-47.3 -33.6,-88 -80.1,-96.9c-34,-6.5 -69,5.4 -92,31.2c0,0 0,0 -0.1,0.1s0,0 -0.1,0.1l-17.8,20c-0.3,0.4 -0.7,0.7 -1,1.1c-4.5,4.5 -10.6,7 -16.9,7s-12.4,-2.5 -16.9,-7z"/>
-</vector>
