+++ app/src/main/res/layout/fragment_profile.xml
@@ -41,8 +41,10 @@
             app:flow_horizontalGap="10dp"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintHorizontal_bias="0.551"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/txt_profile_email">
+            app:layout_constraintTop_toBottomOf="@+id/txt_profile_email"
+            app:layout_constraintVertical_bias="0.6">
 
             <LinearLayout
                 android:id="@+id/logout_btn"
@@ -94,7 +96,7 @@
             android:id="@+id/btn_update_password"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginStart="10dp"
+            android:layout_marginStart="28dp"
             android:layout_marginTop="8dp"
             android:text="@string/profile_btn_password"
             app:layout_constraintStart_toStartOf="parent"
@@ -104,23 +106,21 @@
             android:id="@+id/btn_update"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
+            android:layout_marginTop="8dp"
             android:text="@string/profile_btn_update"
-            android:layout_marginStart="10dp"
-            app:layout_constraintTop_toTopOf="@id/btn_update_password"
-            app:layout_constraintBottom_toBottomOf="@id/btn_update_password"
-            app:layout_constraintStart_toEndOf="@id/btn_update_password"
             app:layout_constraintEnd_toEndOf="parent"
-             />
+            app:layout_constraintHorizontal_bias="0.761"
+            app:layout_constraintStart_toEndOf="@+id/btn_update_password"
+            app:layout_constraintTop_toBottomOf="@+id/txt_profile_email" />
 
         <TextView
             android:id="@+id/txt_profile_email"
-            android:layout_width="match_parent"
+            android:layout_width="350dp"
             android:layout_height="48dp"
             android:layout_marginTop="8dp"
             android:ems="10"
             android:importantForAutofill="no"
             android:textSize="20sp"
-            android:layout_marginHorizontal="15dp"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintHorizontal_bias="0.4"
             app:layout_constraintStart_toStartOf="parent"
@@ -128,10 +128,9 @@
 
         <TextView
             android:id="@+id/textView3"
-            android:layout_width="match_parent"
+            android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginTop="20dp"
-            android:layout_marginHorizontal="15dp"
             android:contentDescription="@string/profile_email_hint"
             android:labelFor="@id/txt_profile_email"
             android:text="@string/profile_email"
@@ -143,27 +142,26 @@
 
         <TextView
             android:id="@+id/txt_profile_username"
-            android:layout_width="match_parent"
-            android:layout_height="40dp"
+            android:layout_width="350dp"
+            android:layout_height="48dp"
             android:layout_marginTop="8dp"
             android:ems="10"
-            android:layout_marginHorizontal="15dp"
             android:importantForAutofill="no"
             android:textSize="20sp"
             app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintHorizontal_bias="0.352"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/textView2" />
 
         <TextView
             android:id="@+id/textView2"
-            android:layout_width="match_parent"
+            android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginTop="30dp"
+            android:layout_marginTop="32dp"
             android:contentDescription="@string/profile_user_hint"
             android:labelFor="@+id/txt_profile_username"
             android:text="@string/profile_username"
             android:textSize="20sp"
-            android:layout_marginHorizontal="15dp"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintHorizontal_bias="0.054"
             app:layout_constraintStart_toStartOf="parent"
